+++ Common/Orders/Order.cs
@@ -31,171 +31,8 @@ namespace QuantConnect.Orders
     /// <summary>
     /// Order struct for placing new trade
     /// </summary>
-    public abstract class Order
+    public abstract class Order: OrderParameters
     {
-        private volatile int _incrementalId;
-        private decimal _quantity;
-        private decimal _price;
-
-        /// <summary>
-        /// Order ID.
-        /// </summary>
-        public int Id { get; internal set; }
-
-        /// <summary>
-        /// Order id to process before processing this order.
-        /// </summary>
-        public int ContingentId { get; internal set; }
-
-        /// <summary>
-        /// Brokerage Id for this order for when the brokerage splits orders into multiple pieces
-        /// </summary>
-        public List<string> BrokerId { get; internal set; }
-
-        /// <summary>
-        /// Symbol of the Asset
-        /// </summary>
-        public Symbol Symbol { get; internal set; }
-
-        /// <summary>
-        /// Price of the Order.
-        /// </summary>
-        public decimal Price
-        {
-            get { return _price; }
-            internal set { _price = value.Normalize(); }
-        }
-
-        /// <summary>
-        /// Currency for the order price
-        /// </summary>
-        public string PriceCurrency { get; internal set; }
-
-        /// <summary>
-        /// Gets the utc time the order was created.
-        /// </summary>
-        public DateTime Time { get; internal set; }
-
-        /// <summary>
-        /// Gets the utc time this order was created. Alias for <see cref="Time"/>
-        /// </summary>
-        public DateTime CreatedTime => Time;
-
-        /// <summary>
-        /// Gets the utc time the last fill was received, or null if no fills have been received
-        /// </summary>
-        [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public DateTime? LastFillTime { get; internal set; }
-
-        /// <summary>
-        /// Gets the utc time this order was last updated, or null if the order has not been updated.
-        /// </summary>
-        [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public DateTime? LastUpdateTime { get; internal set; }
-
-        /// <summary>
-        /// Gets the utc time this order was canceled, or null if the order was not canceled.
-        /// </summary>
-        [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public DateTime? CanceledTime { get; internal set; }
-
-        /// <summary>
-        /// Number of shares to execute.
-        /// </summary>
-        public decimal Quantity
-        {
-            get { return _quantity; }
-            internal set { _quantity = value.Normalize(); }
-        }
-
-        /// <summary>
-        /// Order Type
-        /// </summary>
-        public abstract OrderType Type { get; }
-
-        /// <summary>
-        /// Status of the Order
-        /// </summary>
-        public OrderStatus Status { get; set; }
-
-        /// <summary>
-        /// Order Time In Force
-        /// </summary>
-        [JsonIgnore]
-        public TimeInForce TimeInForce => Properties.TimeInForce;
-
-        /// <summary>
-        /// Tag the order with some custom data
-        /// </summary>
-        [DefaultValue(""), JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string Tag { get; internal set; }
-
-        /// <summary>
-        /// Additional properties of the order
-        /// </summary>
-        public IOrderProperties Properties { get; private set; }
-
-        /// <summary>
-        /// The symbol's security type
-        /// </summary>
-        public SecurityType SecurityType => Symbol.ID.SecurityType;
-
-        /// <summary>
-        /// Order Direction Property based off Quantity.
-        /// </summary>
-        public OrderDirection Direction
-        {
-            get
-            {
-                if (Quantity > 0)
-                {
-                    return OrderDirection.Buy;
-                }
-                if (Quantity < 0)
-                {
-                    return OrderDirection.Sell;
-                }
-                return OrderDirection.Hold;
-            }
-        }
-
-        /// <summary>
-        /// Get the absolute quantity for this order
-        /// </summary>
-        [JsonIgnore]
-        public decimal AbsoluteQuantity => Math.Abs(Quantity);
-
-        /// <summary>
-        /// Gets the executed value of this order. If the order has not yet filled,
-        /// then this will return zero.
-        /// </summary>
-        public decimal Value => Quantity * Price;
-
-        /// <summary>
-        /// Gets the price data at the time the order was submitted
-        /// </summary>
-        public OrderSubmissionData OrderSubmissionData { get; internal set; }
-
-        /// <summary>
-        /// Returns true if the order is a marketable order.
-        /// </summary>
-        public bool IsMarketable
-        {
-            get
-            {
-                if (Type == OrderType.Limit)
-                {
-                    // check if marketable limit order using bid/ask prices
-                    var limitOrder = (LimitOrder)this;
-                    return OrderSubmissionData != null &&
-                           (Direction == OrderDirection.Buy && limitOrder.LimitPrice >= OrderSubmissionData.AskPrice ||
-                            Direction == OrderDirection.Sell && limitOrder.LimitPrice <= OrderSubmissionData.BidPrice);
-                }
-
-                return Type == OrderType.Market;
-            }
-        }
-
         /// <summary>
         /// Added a default constructor for JSON Deserialization:
         /// </summary>
